refactor: move load_external to utils
