+++ src/utils/index.ts
@@ -13,3 +13,4 @@ export * from "./hash"
 export * from "./compatibility"
 export * from "./share"
 export * from "./storage"
+export * from "./load_external"
