+++ src/utils/load_external.ts
